+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed error handling in `internaltopicreader.addOnTransactionCompletedHandler`
+
 ## v3.116.2
 * Fixed grpc connections leak on reused node IP